Change derivation output from key to byte[]
Updating derivation output from SecretKeySpec to generic byte[].

Also updating their tests.
