+++ src/test/java/com/zysn/passwordmanager/model/security/algorithm/derivation/impl/Argon2Test.java
@@ -1,19 +1,16 @@
-package com.zysn.passwordmanager.model.security.algorithm.derivation;
+package com.zysn.passwordmanager.model.security.algorithm.derivation.impl;
 
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 
-import java.util.Arrays;
 import java.util.HashMap;
-import javax.crypto.spec.SecretKeySpec;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import com.zysn.passwordmanager.model.security.algorithm.config.AlgorithmConfig;
 import com.zysn.passwordmanager.model.security.algorithm.derivation.api.KeyDerivationAlgorithm;
-import com.zysn.passwordmanager.model.security.algorithm.derivation.impl.Scrypt;
 
-public class ScryptTest {
+public class Argon2Test {
     private KeyDerivationAlgorithm keyDerivationAlgorithm;
     private char[] source;
     private byte[] salt;
@@ -21,19 +18,23 @@ public class ScryptTest {
 
     @BeforeEach
     void setup() {
-        keyDerivationAlgorithm = new Scrypt();
+        keyDerivationAlgorithm = new Argon2();
         source = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789".toCharArray();
         salt = new byte[] { 72, 92, -108, -126, 80, 92, 114, -96, 13, -93, 69, 96, -89, -25, 34, -102, 77, -20, -8, -41,
                 92, 87, 17, 35, 75, -11, -87, -87, -54, 22, 110, 8 };
 
-        algorithmConfig = new AlgorithmConfig("Key Derivation Algorithm", "Scrypt", salt, new HashMap<>());
+        algorithmConfig = new AlgorithmConfig("Key Derivation Algorithm", "Argon2", salt, new HashMap<>());
 
-        String iterationsName = "cost_factor";
-        String iterationsValue = "16384";
+        String variantName = "variant";
+        String variantValue = "argon2id";
+        algorithmConfig.addNewParameter(variantName, variantValue);
+
+        String iterationsName = "iterations";
+        String iterationsValue = "3";
         algorithmConfig.addNewParameter(iterationsName, iterationsValue);
 
-        String memoryCostName = "block_size";
-        String memoryCostValue = "8";
+        String memoryCostName = "memory_cost";
+        String memoryCostValue = "8192";
         algorithmConfig.addNewParameter(memoryCostName, memoryCostValue);
 
         String parallelismName = "parallelism";
@@ -48,13 +49,11 @@ public class ScryptTest {
 
     @Test
     void testDeriveKey() {
-        SecretKeySpec masterKey = keyDerivationAlgorithm.deriveKey(source, algorithmConfig);
+        byte[] actualMasterKey = keyDerivationAlgorithm.deriveKey(source, algorithmConfig);
 
-        byte[] masterKeyByte = new byte[] { -107, 29, 119, 84, -17, 116, 114, 121, -127, -37, 72, 111, -77, -106, -110,
-                -1 };
+        byte[] expectedMasterKey = new byte[] { 92, 124, 24, -128, 14, -101, -19, -96, -49, -59, 10, -77, -111, -39, -9,
+                -91 };
 
-        assertArrayEquals(masterKeyByte, masterKey.getEncoded(),
-                "Obtained value is: " + Arrays.toString(masterKey.getEncoded()) + " but the real value should be: "
-                        + masterKeyByte);
+        assertArrayEquals(expectedMasterKey, actualMasterKey, "The Master Key arrays do not match.");
     }
 }
