+++ src/test/java/com/zysn/passwordmanager/model/security/algorithm/derivation/patterns/KeyDerivationAlgorithmFactoryTest.java
@@ -1,5 +1,6 @@
-package com.zysn.passwordmanager.model.security.algorithm.derivation;
+package com.zysn.passwordmanager.model.security.algorithm.derivation.patterns;
 
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.BeforeEach;
@@ -7,7 +8,6 @@ import org.junit.jupiter.api.Test;
 
 import com.zysn.passwordmanager.model.security.algorithm.derivation.api.KeyDerivationAlgorithm;
 import com.zysn.passwordmanager.model.security.algorithm.derivation.impl.Argon2;
-import com.zysn.passwordmanager.model.security.algorithm.derivation.patterns.KeyDerivationAlgorithmFactory;
 
 public class KeyDerivationAlgorithmFactoryTest {
 
@@ -21,5 +21,14 @@ public class KeyDerivationAlgorithmFactoryTest {
         KeyDerivationAlgorithm keyDerivationAlgorithm = KeyDerivationAlgorithmFactory.createAlgorithm(algorithmName);
 
         assertTrue(keyDerivationAlgorithm instanceof Argon2, "The object is not an istance of 'Argon2'");
+    }
+
+    @Test
+    void testCreateWrongAlgorithm() {
+        String algorithmName = "Argon-1";
+
+        assertThrows(IllegalArgumentException.class, () -> {
+            KeyDerivationAlgorithmFactory.createAlgorithm(algorithmName);
+        });
     }
 }
